Dieser PR dient erstmal nur einer kurzen Diskussion zur Grundproblematik und der Frage ob und wie man das angehen kann.

Ich habe die Performance des Einnahmen/Ausgaben-View nochmal genauer angeschaut. Die Probleme fallen insb. bei monatlicher Gruppierung auf (für mehrere Konten wird es dann auch nicht besser). Eine Auffälligkeit war, dass auch bei Konten ohne Umsätze (z.B. gekündigt, aber noch in einer Auswertungs-Gruppe) der Aufbau der Daten extrem lange dauert.

Es sind insb. die Anfangs- und Endsaldo-Ermittlungen, die lange dauern. Je nachdem ob der Zeitraum am Anfang oder Ende des Kontolebenszyklus liegt Faktor bei 10 (Buchungssummen 300ms, Saldo 4s; das Konto hat über 1000 Buchungen). Die Gründe sind im Code ja auch schon beschrieben - dediziertes SQL nicht möglich. In der aktuellen Implementierung führt das aber dazu, dass sämtliche Umsätze 12x (Analysezeitraum 1 Jahr, monatliche Gruppierung) instantiiert werden. Da wäre es bei einer Gruppierung effizienter, sich alle Umsätze des Kontos einmal zu holen und alles selbst auszurechnen anstatt 12x(4+X) SQL-Statements über das Konto-Util abzusetzen.

Die im PR verwendete Vorgängerzeitraums-Map-Änderung beschleunigt erstmal nur den Fall, dass es für ein oder mehrere Konten keine Umsätze in einem Zeitraum gibt, ist aber nur Symptombekämpfung (Vermeidung unnötiger langsamer Zugriffe) und hilft nicht für den Standardfall.
Final die beste Lösung wäre meiner Ansicht nach, wenn man beim DbIterator nicht nur ein limit angeben könnte, sondern auch eine Page-Size - es werden immer nur so viele Objekte instantiiert und wenn die Page zuende ist, wird beim nächsten next() wieder an die Datenbank gegangen.
Wenn das prinzipiell nicht machbar ist, wäre mein Vorschlag, das Zusammenbauen der EinnahmenAusgaben-Objekte für den gruppierten Fall komplett in Java zu machen - es werden ohnehin praktisch alle Umsätze des Kontos instantiiert (für Ermittlung des Anfangssaldo und Endsaldo müssen auch die Umsätze vor und nach dem Zeitraum geholt werden).

Das Einführen des aktuell auskommentierten Limits im Konto-Util bringt schon eine drastische Beschleunigung. Leider kennt man das korrekte Limit nicht - zu klein und das Ergebnis ist falsch, zu groß und die Performanceverbesserung ist suboptimal. Hinweis selbst bei Limit 1000, war die Beschleunigung bei mir schon deutlich spürbar: Saldoberechnung 300ms statt 3 s.

Wie siehst Du das?
* paging umsetzbar
* vieleicht doch irgendwie Flag-Auswertung im SQL möglich (Division/Rest-Berechungn auf Spalte)
* Berechnung im Java bei Gruppierung
* limit 1000 OK
* Es gibt gar kein Problem.