@nittka 2020-01-18
Bei dem Versuch das Paging in Java nachzubauen (limit setzen und wenn es Umsätze gab - aber keine ohne Vormerkflag - dann den Filter anpassen und weitersuchen), ist mir aufgefallen, dass in der Methode für die Summenberechnung durchaus sql-seitig auf das "gebucht"-Flag selektiert wird. Es gibt dort das Methodenargument "onlyBooked". Kann man in den Saldomethoden nicht einen analogen Filter hinzufügen, das Limit auf 1 setzen, da dann doch der Saldo des ersten gefundenen Umsatzes verwendbar sein müsste?