@nittka 2020-01-18
Statt des Nachbauens von Paging (sehr fehleranfällig) habe ich mal ein iteratives Erweitern des Suchfensters implementiert. Die Optimierung, gar nicht erst Salden aus Umsätzen zu ermitteln, wenn es keine gab, habe ich drin gelassen. Die Zeiten für das Ermitteln der Daten für die Auswertung gehen stark nach unten.
Statt der vielen boolean-Argumente wäre ein Parameter-Datenobjekt besser für die Lesbarkeit, aber ich betrachte den Ansatz zunächst ohnehin als Proof of Concept.