@willuhn 2020-01-20
Danke für dein Patch. Ich habe ehrlich gesagt große Probleme, deinen Code-Änderungen zu folgen. Ich würde als Lösung des Performance-Problems aber eher vorschlagen, die Berechnungen komplet in Java zu machen. Also pauschal alle Umsätze der gewählten Konten in dem Zeitraum zu laden, die Liste zwischenzuspeichern und dann nur noch auf dieser Liste arbeiten.