@nittka 2020-01-20
Sorry, habe aus Versehen ein Review gestartet. Ich wollte nur kommentieren...
Ich habe mal versucht, kurz die Motivation für die gemachten Änderungen dranzuschreiben.
Ich werde in der nächsten Zeit mal versuchen, Deinen Vorschlag, die gesamte Berechnung in JAva zu machen umzusetzen und die Zeiten zu vergleichen.

Es wird allerdings nicht immer ausreichen, nur die Transaktionen aus dem Suchzeitraum zu betrachten - eben genau wegen der Anfangs- und Endsalden. Gibt es nämlich im Zeitraum keine Transaktionen, muss man wieder davor (und danach?) zu suchen. Aber in diesem Fall reicht dann wohl der Anfangssaldo, denn wenn es keine Transaktion gibt, muss der Anfangssaldo gleich dem Endsaldo sein. Außer man will solche Fälle behandeln, wo der der Endsaldo einer Transaktion nicht dem Anfangssaldo der Nachfolgetransaktion entspricht (Löschen von Transaktionen). Ohne diese Sonderbehandlung kann man nämlich keine Differenzen aufdecken.
Aber das bedeutet wieder, 2x sämtliche Transaktionen zu instantiieren, wenn es kein Paging gibt.

Ich melde mich dann mal mit einem komplett neuen Branch für diese Änderungen zurück. Danke für den Hinweis, in welche Richtung es gehen soll.