@willuhn 2020-01-21
Um die Queries für die Ermittlung der Start- und End-Salden kommt man sicher nicht drum rum. Aber der ganze Rest sollte sich mit einem Query erledigen lassen.