@nittka 2020-01-25
Neuer Pull-Request mit Berechnung in Java ist erstellt.